+++ plugins/scrcpymac-phone-agent/server/phone_agent/recipes/__init__.py
@@ -1,3 +1,0 @@
-from phone_agent.recipes.wechat import send_message
-
-__all__ = ["send_message"]